requirements.txt
Django==1.6.1
Pillow==2.3.0
South==0.8.4
django-ckeditor-updated==4.2.6
django-taggit==0.11.2
wsgiref==0.1.2
